this has a different type in 64 bit systems, so we need to match with other values
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@34847 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: e7533e668423ddba97d36625d942af520c267407

+++ src/router/libutils/libutils/linux_timer.c
@@ -84,7 +84,7 @@ typedef void (*event_callback_t) (timer_t, int);
 	if ((t)->tv_sec == 0) (t)->tv_usec=ROUNDUP((t)->tv_usec, g); \
     } while (0)
 
-typedef long uclock_t;
+typedef int uclock_t;
 
 #define TFLAG_NONE	0
 #define TFLAG_CANCELLED	(1<<0)
